+++ docs/source/about/release-notes.rst
@@ -12,6 +12,7 @@ v1.2.0: Unreleased
 * Support configuring FileAppender#bufferSize `#1951 <https://github.com/dropwizard/dropwizard/pull/1951>`_
 * Improve error handling of `@FormParam` resources `#1982 <https://github.com/dropwizard/dropwizard/pull/1982>`_
 * Support Dropwizard applications without logback `#1900 <https://github.com/dropwizard/dropwizard/pull/1900>`_
+* Replace deprecated SizeAndTimeBasedFNATP with SizeAndTimeBasedRollingPolicy `#1900 <https://github.com/dropwizard/dropwizard/pull/2100>`_
 
 .. _rel-1.1.0: